Move `ExpectedParticipantData` to common.
