+++ carp.protocols.core/src/commonMain/kotlin/dk/cachet/carp/protocols/application/ProtocolServiceHost.kt
@@ -1,8 +1,8 @@
 package dk.cachet.carp.protocols.application
 
 import dk.cachet.carp.common.application.UUID
+import dk.cachet.carp.common.application.users.ExpectedParticipantData
 import dk.cachet.carp.common.application.users.ParticipantAttribute
-import dk.cachet.carp.protocols.application.users.ExpectedParticipantData
 import dk.cachet.carp.protocols.domain.StudyProtocol
 import dk.cachet.carp.protocols.domain.StudyProtocolRepository
 import kotlinx.datetime.Clock
