+++ carp.protocols.core/src/commonMain/kotlin/dk/cachet/carp/protocols/domain/configuration/ParticipantConfiguration.kt
@@ -1,8 +1,8 @@
 package dk.cachet.carp.protocols.domain.configuration
 
+import dk.cachet.carp.common.application.users.ExpectedParticipantData
 import dk.cachet.carp.common.application.users.ParticipantAttribute
 import dk.cachet.carp.common.application.users.ParticipantRole
-import dk.cachet.carp.protocols.application.users.ExpectedParticipantData
 
 
 /**
