+++ carp.protocols.core/src/commonTest/kotlin/dk/cachet/carp/protocols/domain/configuration/ParticipantConfigurationTest.kt
@@ -2,10 +2,10 @@ package dk.cachet.carp.protocols.domain.configuration
 
 import dk.cachet.carp.common.application.data.input.InputDataType
 import dk.cachet.carp.common.application.data.input.elements.Text
+import dk.cachet.carp.common.application.users.ExpectedParticipantData
 import dk.cachet.carp.common.application.users.ParticipantAttribute
 import dk.cachet.carp.common.application.users.ParticipantRole
 import dk.cachet.carp.common.infrastructure.test.createTestJSON
-import dk.cachet.carp.protocols.application.users.ExpectedParticipantData
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import kotlin.test.*
