+++ rpc/schemas/common/users/ExpectedParticipantData.json
@@ -2,7 +2,7 @@
   "$schema": "https://json-schema.org/draft/2019-09/schema",
   "type": "object",
   "properties": {
-    "attribute": { "$ref": "../../common/users/ParticipantAttribute.json" },
+    "attribute": { "$ref": "ParticipantAttribute.json"},
     "inputBy": { "$ref": "#/$defs/InputBy" }
   },
   "required": [ "attribute" ],
@@ -20,7 +20,7 @@
       "$anchor": "InputByAnyone",
       "type": "object",
   	  "properties": {
-        "__type": { "const": "dk.cachet.carp.protocols.application.users.ExpectedParticipantData.InputBy.Anyone" }
+        "__type": { "const": "dk.cachet.carp.common.application.users.ExpectedParticipantData.InputBy.Anyone" }
       },
       "required": [ "__type" ]
     },
@@ -28,7 +28,7 @@
       "$anchor": "InputByRoles",
       "type": "object",
       "properties": {
-        "__type": { "const": "dk.cachet.carp.protocols.application.users.ExpectedParticipantData.InputBy.Roles" },
+        "__type": { "const": "dk.cachet.carp.common.application.users.ExpectedParticipantData.InputBy.Roles" },
         "roleNames": {
           "type": "array",
           "items": "string"
