+++ typescript-declarations/tests/carp.common.ts
@@ -5,6 +5,9 @@ import { kotlin } from 'kotlin'
 import toSet = kotlin.collections.toSet_us0mfu$;
 import Duration = kotlin.time.Duration;
 
+import { kotlinx } from 'kotlinx-serialization-kotlinx-serialization-json-js-legacy'
+import Json = kotlinx.serialization.json.Json
+
 import { dk } from 'carp.core-kotlin-carp.common'
 import EmailAddress = dk.cachet.carp.common.application.EmailAddress;
 import NamespacedId = dk.cachet.carp.common.application.NamespacedId;
@@ -27,10 +30,14 @@ import Text = dk.cachet.carp.common.application.data.input.elements.Text;
 import AccountIdentity = dk.cachet.carp.common.application.users.AccountIdentity;
 import EmailAccountIdentity = dk.cachet.carp.common.application.users.EmailAccountIdentity;
 import ParticipantAttribute = dk.cachet.carp.common.application.users.ParticipantAttribute;
+import ExpectedParticipantData = dk.cachet.carp.common.application.users.ExpectedParticipantData
+import InputBy = dk.cachet.carp.common.application.users.ExpectedParticipantData.InputBy
+import Roles = dk.cachet.carp.common.application.users.ExpectedParticipantData.InputBy.Roles
 import Username = dk.cachet.carp.common.application.users.Username;
 import UsernameAccountIdentity = dk.cachet.carp.common.application.users.UsernameAccountIdentity;
 import emailAccountIdentityFromString = dk.cachet.carp.common.application.users.EmailAccountIdentity_init_61zpoe$;
 import ApiVersion = dk.cachet.carp.common.application.services.ApiVersion
+import createDefaultJSON = dk.cachet.carp.common.infrastructure.serialization.createDefaultJSON_18xi4u$
 
 import { dk as ddk } from 'carp.core-kotlin-carp.deployments.core'
 import DeploymentServiceRequest = ddk.cachet.carp.deployments.infrastructure.DeploymentServiceRequest
@@ -79,6 +86,10 @@ describe( "carp.common", () => {
             UsernameAccountIdentity.Companion,
             [ "ParticipantAttribute", new ParticipantAttribute.DefaultParticipantAttribute( new NamespacedId( "namespace", "type" ) ) ],
             ParticipantAttribute.Companion,
+            new ExpectedParticipantData( new ParticipantAttribute.DefaultParticipantAttribute( new NamespacedId( "namespace", "type" ) ) ),
+            ExpectedParticipantData.Companion,
+            InputBy.Companion,
+            InputBy.Anyone,
             new ApiVersion( 1, 0 ),
             [ "ApplicationServiceRequest", new DeploymentServiceRequest.GetStudyDeploymentStatus( UUID.Companion.randomUUID() ) ]
         ]
@@ -127,6 +138,20 @@ describe( "carp.common", () => {
         it( "inputToData works", () => {
             const data = attribute.inputToData_etkzhw$( CarpInputDataTypes, "Steven" )
             expect( data ).is.not.undefined
+        } )
+    } )
+
+    describe( "ExpectedParticipantData", () => {
+        it( "can serialize polymorphic InputBy", () => {
+            const expectedData = new ExpectedParticipantData(
+                new ParticipantAttribute.DefaultParticipantAttribute( new NamespacedId( "namespace", "type" ) ),
+                new Roles( toSet( [ "Roles are added" ] ) )
+            )
+
+            const json: Json = createDefaultJSON()
+            const serializer = ExpectedParticipantData.Companion.serializer()
+            const serialized = json.encodeToString_tf03ej$( serializer, expectedData )
+            expect( serialized ).has.string( "Roles are added" )
         } )
     } )
 } )
