+++ typescript-declarations/tests/carp.protocols.core.ts
@@ -8,17 +8,12 @@ import { kotlinx } from 'kotlinx-serialization-kotlinx-serialization-json-js-leg
 import Json = kotlinx.serialization.json.Json
 
 import { dk as cdk } from 'carp.core-kotlin-carp.common'
-import NamespacedId = cdk.cachet.carp.common.application.NamespacedId
 import UUID = cdk.cachet.carp.common.application.UUID
-import ParticipantAttribute = cdk.cachet.carp.common.application.users.ParticipantAttribute
 import createDefaultJSON = cdk.cachet.carp.common.infrastructure.serialization.createDefaultJSON_18xi4u$
 
 import { dk } from 'carp.core-kotlin-carp.protocols.core'
 import ProtocolVersion = dk.cachet.carp.protocols.application.ProtocolVersion
 import StudyProtocolSnapshot = dk.cachet.carp.protocols.application.StudyProtocolSnapshot
-import ExpectedParticipantData = dk.cachet.carp.protocols.application.users.ExpectedParticipantData
-import InputBy = dk.cachet.carp.protocols.application.users.ExpectedParticipantData.InputBy
-import Roles = dk.cachet.carp.protocols.application.users.ExpectedParticipantData.InputBy.Roles
 import ProtocolFactoryServiceRequest = dk.cachet.carp.protocols.infrastructure.ProtocolFactoryServiceRequest
 import ProtocolServiceRequest = dk.cachet.carp.protocols.infrastructure.ProtocolServiceRequest
 
@@ -37,10 +32,6 @@ describe( "carp.protocols.core", () => {
             ProtocolVersion.Companion,
             studyProtocolSnapshot,
             StudyProtocolSnapshot.Companion,
-            new ExpectedParticipantData( new ParticipantAttribute.DefaultParticipantAttribute( new NamespacedId( "namespace", "type" ) ) ),
-            ExpectedParticipantData.Companion,
-            InputBy.Companion,
-            InputBy.Anyone,
             [ "ProtocolServiceRequest", new ProtocolServiceRequest.GetAllForOwner( UUID.Companion.randomUUID() ) ],
             [ "ProtocolFactoryServiceRequest", new ProtocolFactoryServiceRequest.CreateCustomProtocol( UUID.Companion.randomUUID(), "", "" ) ]
         ]
@@ -55,20 +46,6 @@ describe( "carp.protocols.core", () => {
             const serializer = StudyProtocolSnapshot.Companion.serializer()
             const parsed = json.decodeFromString_awif5v$( serializer, serializedSnapshot )
             expect( parsed ).is.instanceOf( StudyProtocolSnapshot )
-        } )
-    } )
-
-    describe( "ExpectedParticipantData", () => {
-        it( "can serialize polymorphic InputBy", () => {
-            const expectedData = new ExpectedParticipantData(
-                new ParticipantAttribute.DefaultParticipantAttribute( new NamespacedId( "namespace", "type" ) ),
-                new Roles( toSet( [ "Roles are added" ] ) )
-            )
-
-            const json: Json = createDefaultJSON()
-            const serializer = ExpectedParticipantData.Companion.serializer()
-            const serialized = json.encodeToString_tf03ej$( serializer, expectedData )
-            expect( serialized ).has.string( "Roles are added" )
         } )
     } )
